Expense Project Creation And CSVService.py writing
I have started a quick project according to a SimpliLearn class that I
started taking to further my AI development, as I have not had the
chance to do applied generative AI material in college and the
curriculum states that I won't

This first project is to get our feet wet with Python. Good for me, I'm
already decently knowledgeable in Python, and have challenged myself to
write well documented and clean code for this exercise

There are 4 files present: main, CSVService, StateManager, and
UIPrompts, with a 5th coming later, simply named data.csv. The goal of
the project is simple; make a terminal expense tracker that can read and
write a csv file that stores different information... more on that to
come later. I've uploaded the pdf of the requirements to look at.

I've completely outlined this project in my notes (as much as I can) and
have challenged myself to upload one file at a time and not do any
testing until I get to the final product. That way I can really see if I
know how to formulate projects structurally like I think I do, or if I'm
just full of it. Let's hope the former!