Adding Basic UI Prompts and Slight Manager Change
The basic UI prompts that I know are going to be repeated a lot
have been added. I'll add the rest of them soon enough, since it would
make sense for all of them to be in the same file, but for right now
it'll be ok

Some slight Manager changes: I had erroneously left a returning int when
I want to return a string. I also changed the budget to an infinite
float since that's a bit more representative of what it will look like

Interestingly, I haven't added in the get/set for the budget or the get
for the totalSpent. Not that I necessarily need to, but following good
coding practice is pretty essential right now.