Created the UI
Here's how it goes: The UI is a multi-inherited class of the State and
CSVService that is meant to be the output of the two. That's it, that's
how it's built. 3 classes, one of them inheriting the other two.

I thought it was the best way to go about this... seemed pretty natural
structure wise. Makes it pretty easy to change later down the line...
However, I might need to add in an interface/class for the expense,
since as of right now that's not entirely easy to change... there's a
couple places that would have to change if someone wanted to add on to
the code

However, as of right now, this seems pretty solid

I have not done ANY testing yet, because I wanted a commit before and
after I started testing, so that I can create branches to show how I
messed up, the things that need to be fixed, etc. But for now, I think I
did a pretty good job