Changing the nameless dictionary to ExpenseClass
This branch is meant to introduce a new class
for expenses. Before I had the forgone conclusion
that nameless dictionaries were the way to go,
but over the past day I had the realization that
a class would be a better approach for making the
code more maintainable and easier to changes/read.