Refactor load plus UI prompts
Previously, I wanted to make the load functionality "dynamic"
by giving the users an option on whether to append
the loaded data to the existing data or overwrite it.

There is not a load feature in this entire world
that attempts to do such a thing. If you want to load
data, you load it and replace it. That's it.

In most loads, there are SOME things that are appended,
but that is usually metadata that is not current or
replacing visualized data.

So no, it doesn't need to be as complicated as I had made
it out to be.

This commit also refactors the UI prompts a bit
to accomodate for that change