Refactoring Validations
This commit refactors the validation logic in the expense tracker application.
Previously, validation checks were being handeled in teh functions that processed user input.
I read recently in the Pragmatic Programmer that the best practice is to keep validation logic separate from the processing logic.

For example: I shouldn't be passing potentially bad data into the create_expense function
The purpose of that is to create an expense... not to validate it.