Fixing initialization "issues"
There actually isn't inherently a problem with initialization...
The code works fine.
However, there is an issue in how expenses were being created
My original code made assumptions all over the place about
how many parameters were being passed to the Expense class.

That should never happen, since the needs of a program
can change all the time
Since the expense class is just an encapsulation of
a dictionary, I can just pass it the full dictionary
and initialize it in one place: where I create it.
Past that, nothing else should matter

My code is now a lot easier to change if in the future someone wanted to
add other sorts of data as an input, such as another date
or description column

The only thing that I couldn't make to dynamically set was the
table that is outputted in option 2. Tables are....
far too complicated for me to try and make in
a thirty minute debug span. In fact,
a whole project could just be a table. Those things
are touch. So, that willstay as it is.